Ensure ABS calls work on difference of unsigned values
guint64 - guint64 is of type guint64, so ABS doesn't work on it.
Cast to gint64 to make sure this works.
(cherry picked from commit 1f8f54f4777343dda87043e444a24c3f86e90439)

+++ libnautilus-private/nautilus-file-operations.c
@@ -1396,7 +1396,7 @@ report_delete_progress (CommonJob *job,
 
 	now = g_thread_gettime ();
 	if (transfer_info->last_report_time != 0 &&
-	    ABS (transfer_info->last_report_time - now) < 100 * NSEC_PER_MSEC) {
+	    ABS ((gint64)(transfer_info->last_report_time - now)) < 100 * NSEC_PER_MSEC) {
 		return;
 	}
 	transfer_info->last_report_time = now;
@@ -2820,7 +2820,7 @@ report_copy_progress (CopyMoveJob *copy_job,
 	now = g_thread_gettime ();
 	
 	if (transfer_info->last_report_time != 0 &&
-	    ABS (transfer_info->last_report_time - now) < 100 * NSEC_PER_MSEC) {
+	    ABS ((gint64)(transfer_info->last_report_time - now)) < 100 * NSEC_PER_MSEC) {
 		return;
 	}
 	transfer_info->last_report_time = now;
